+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Info menu something not showing up
+- Command argument handling
 
 ## [0.30] - 2020-04-17
 ### Changed